Move generic command execution code into shared CLI library.
