+++ cardano-wallet.cabal
@@ -38,9 +38,7 @@ executable cardano-wallet
       -Werror
   build-depends:
       base
-    , aeson-pretty
     , async
-    , bytestring
     , cardano-wallet-cli
     , cardano-wallet-core
     , cardano-wallet-http-bridge
@@ -56,7 +54,6 @@ executable cardano-wallet
     , process
     , servant-client
     , servant-client-core
-    , servant-server
     , text
     , text-class
     , warp
