+++ lib/cli/cardano-wallet-cli.cabal
@@ -30,6 +30,7 @@ library
       -Werror
   build-depends:
       aeson
+    , aeson-pretty
     , ansi-terminal
     , base
     , bytestring
@@ -39,6 +40,9 @@ library
     , heredoc
     , http-client
     , iohk-monitoring
+    , servant-client
+    , servant-client-core
+    , servant-server
     , text
     , text-class
   hs-source-dirs:
